--- src/codemodder/project_analysis/file_parsers/utils.py
@@ -1,7 +1,0 @@
-import libcst as cst
-
-
-def clean_simplestring(node: cst.SimpleString | str) -> str:
-    if isinstance(node, str):
-        return node.strip('"')
-    return node.raw_value
